moving setupLogging() to profile, and updating all apps to look in new location
